cmake: Adjust libcaffeine to new CPack packages
Improves the libcaffeine find module to search for both a CPack release and a normal project directory, to support both released and in-development versions of the project.

The CPack search will require a CaffeineConfig.cmake which is immediately included, and defines various variables and a libcaffeine target for linking. The Project search works similar, but only links requires a single Release configuration to be built, but can only differentiate between Release and Debug configurations due to the restrictions placed on additional_install_files.