cmake: Adjust libcaffeine to new CPack packages
This updated libcaffeine 'Find Module' file is able to deal with both in-project and CPacked releases of libcaffeine. The configuration defines a target called libcaffeine which should be used to link against, instead of using the variables directly.

The CPack version will search for a properly packaged CPack install of libcaffeine in either LIBCAFFEINE_DIR(32|64) or DepsPath(32|64), while the project version will search build(32|64) and build/<32|64> for compiled files. At least Debug and one Release configuration has to be built for the project based one, and all configurations have to be built for the CPack based one.

Due to the use of find_path and find_file, variables may have to be deleted first from the CMakeCache in order for them to be used by a CPacked archive.